Move format_signed to utils
